pyproject.toml
[project]
name = "vertex"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform>=1.138.0",
    "google-generativeai",
    "pydantic",
    "absl-py",
    "mcp[cli]>=1.21.2",
    "google-cloud-storage",
    "pandas",
    "matplotlib",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = ["ruff~=0.5.5", "pyright", "pytest", "ipython", "pre-commit"]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "W",  # pycodestyle warnings
    "I",  # isort
    "UP", # pyupgrade
    "D",  # pydocstyle
    "N",  # pep8-naming
    "PL", # Pylint
]
ignore = [
    "D100",    # Missing docstring in public module (module docstrings for tests are not required, and for others, can be added as needed)
    "D101",    # Missing docstring in public class (class docstrings are important, but for very simple classes, it might be deferred)
    "D102",    # Missing docstring in public method (methods with obvious purpose or trivial implementation might not need a docstring)
    "D103",    # Missing docstring in public function (functions with obvious purpose or trivial implementation might not need a docstring)
    "D104",    # Missing docstring in public package (similar to module docstrings)
    "D105",    # Missing docstring in magic method (magic methods like __init__ often have self-explanatory purposes)
    "D107",    # Missing docstring in public __init__ (explicitly for __init__, often self-explanatory)
    "PLR0913", # Too many arguments in function definition (common for CLI tools)
    "E501",    # Line too long (auto-formatter not fixing for specific docstrings, and style guide allows exceptions)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.packages.find]
where = ["src"]
